More cleanup in preparation for new authentication flow
